Abort drop tablespace on error, delete files on commit.
We add information to the TwoPhaseHeaderFile about which tablespace we're
attempting to drop. We also schedule a tablespace for deletion (similar
to CREATE TABLESPACE) on commit, which waits until the transaction
is being committed to perform the filesystem deletion.

We also modified the distributed_commit record to pass the tablespace to
be dropped to the standby. And, we modified the abort record, to help
back out of a drop on the mirrors.