Map filespaces in tablespaces SQL command in pg_dumpall
Implement pg_dumpall processing of FILESPACE and TABLESPACE in 5.X to
map these objects to 6.X.

To do that, a DDL describing 5.X TABLESPACE object in 6.X style is
formed. Actual locations used are the same as old ones, but with
'GP6/<tablespace_oid>' appended.

Note that this solution does not work during pg_upgrade. In Greenplum,
pg_upgrade runs only master segment, executes schema restore, and
copies master data directory to segments; this does not allow to
properly create TABLESPACEs on segments. Thus, the same location
(defined for master segment) would be used on all segments, despite
the presence of particular locations in TABLESPACE DDL.

Co-authored-by: Ivan Leskin <leskin.in@arenadata.io>
Co-authored-by: Soumyadeep Chakraborty <sochakraborty@pivotal.io>