Upgrade tablespaces
The administrator will be responsible for generating a CSV file to be
passed to pg_upgrade containing the dbid, tablespace oid, and
filespace path for each tablespace in the old cluster. We can then use
that file to perform the filespace to tablespace mapping.

We issue a create tablespace ... location on the master during the
upgrade using the same location as the master's filespace.

There is some work that needs to be done by an external upgrade
process (maybe gpupgrade) to copy data from the master node to the
segments, including the master tablespace directories.

* pass file to pg_upgrade containing old cluster's tablespace locations, oid, and dbid.
* output tablespace creation w/ location in pg_dumpall
* parse the given CSV to determine old tablespace location for mapping
  relfiles